requirements.txt
Django==5.2.4
python-dotenv==1.1.1
dj-database-url==3.0.1
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0